Fix label dtype in _TokenizedDataset causing loss stuck at 1.35
torch.tensor(self.labels[idx]) infers dtype from the input. When the
pandas label column is float64 (after dataset merges with NaN coercion),
the tensor becomes float32. CrossEntropyLoss requires torch.long — with
float labels it computes a malformed loss that stays stuck near 1.35
(2*log(2)) across all epochs, so the model never learns.

Fix: explicit int() cast + dtype=torch.long on the label tensor.

+++ pidm/detection/classifier.py
@@ -33,7 +33,7 @@ class _TokenizedDataset(TorchDataset):
 
     def __getitem__(self, idx):
         item = {k: torch.tensor(v[idx]) for k, v in self.encodings.items()}
-        item["labels"] = torch.tensor(self.labels[idx])
+        item["labels"] = torch.tensor(int(self.labels[idx]), dtype=torch.long)
         return item
